working dpc test + timer adapter
