+++ genode_env/phantom_env.h
@@ -3,10 +3,10 @@
 
 #include <base/component.h>
 #include <base/heap.h>
-
 #include "disk_backend.h"
 #include "phantom_threads.h"
 #include "phantom_vmem.h"
+#include "phantom_timer.h"
 
 namespace Phantom
 {
@@ -20,6 +20,7 @@ namespace Phantom
         Phantom::Disk_backend _disk{_env, _heap};
         Phantom::PhantomThreadsRepo _threads_repo{_env, _heap};
         Phantom::Vmem_adapter _vmem_adapter{_env};
+        Phantom::Timer_adapter _timer_adapter{_env};
 
         Main(Env &env) : _env(env)
         {
