+++ phantom/isomem/genode_timer.c
@@ -1,6 +1,6 @@
-#include <kernel/board.h>
-#include "genode_misc.h"
+// #include <kernel/board.h>
+// #include "genode_misc.h"
 
-void board_init_kernel_timer(void){
-    _stub_print();
-}
+// void board_init_kernel_timer(void){
+//     _stub_print();
+// }
