+++ phantom/isomem/test_switch.c
@@ -184,10 +184,10 @@ void run_test( const char *test_name, const char *test_parm )
 
 
     // These are long
-    /* //   TODO : Uncomment
+    //   TODO : Uncomment
     TEST(dpc);
     TEST(timed_call);
-    */
+    
 
     // must test after timed calls for it depends on them
     // TEST(ports);
